+++ src/providers/AppsProvider.js
@@ -1,18 +1,12 @@
 /* eslint-disable semi,@typescript-eslint/no-var-requires */
 const vscode = require('vscode');
-
 const App = require('../tree/App');
 const Group = require('../tree/Group')
 const Item = require('../tree/Item')
 const Code = require('../tree/Code')
 const Core = require('../Core');
-const Meta = require('../Meta');
-
-const path = require('path')
-const { mkdir } = require('fs/promises');
-const download = require('image-downloader')
-const asyncfile = require('async-file')
 const camelCase = require('lodash.camelcase');
+const { downloadAndStoreAppIcon } = require('../libs/app-icon');
 
 class AppsProvider /* implements vscode.TreeDataProvider<Dependency> */ {
 	constructor(_authorization, _environment, _DIR, _admin) {
@@ -57,52 +51,11 @@ class AppsProvider /* implements vscode.TreeDataProvider<Dependency> */ {
 					break;
 			}
 			if (response === undefined) { return }
-			let iconDir = path.join(this._DIR, "icons")
-			await mkdir(iconDir, { recursive: true });
-			let apps = response.map(async (app) => {
-				let iconVersion = 1
-				let dest = path.join(iconDir, `${app.name}.${iconVersion}.png`)
-				while (await asyncfile.exists(`${dest}.old`)) {
-					iconVersion++
-					dest = path.join(iconDir, `${app.name}.${iconVersion}.png`)
-				}
-				if (!await asyncfile.exists(dest)) {
-					try {
-						await download.image({
-							headers: {
-								"Authorization": this._authorization,
-								'x-imt-apps-sdk-version': Meta.version
-							},
-							url: (() => {
-								switch (this._environment.version) {
-									case 2:
-										return `${this._baseUrl}/sdk/apps/${app.name}/${app.version}/icon/512`
-									case 1:
-									default:
-										return `${this._baseUrl}/app/${app.name}/${app.version}/icon/512`
-								}
-							})(),
-							dest: dest
-						})
-						await Core.invertPngAsync(dest);
-					}
-					catch (err) {
-						if (err != undefined) {
-							iconVersion = 0
-						}
-					}
-				}
-				if (app.public) {
-					if (await asyncfile.exists(dest) && !await asyncfile.exists(`${dest.slice(0, -4)}.public.png`)) {
-						await Core.generatePublicIcon(dest);
-					}
-					if (await asyncfile.exists(`${dest.slice(0, -4)}.dark.png`) && !await asyncfile.exists(`${dest.slice(0, -4)}.dark.public.png`)) {
-						await Core.generatePublicIcon(`${dest.slice(0, -4)}.dark.png`);
-					}
-				}
-				return new App(app.name, app.label, app.description, app.version, app.public, app.approved, iconDir, app.theme, app.changes, iconVersion)
-			})
-			apps = await Promise.all(apps)
+
+			const apps = await Promise.all(response.map(async (app) => {
+				const iconVersion = await downloadAndStoreAppIcon(app, this._baseUrl, this._authorization, this._environment, false);
+				return new App(app.name, app.label, app.description, app.version, app.public, app.approved, app.theme, app.changes, iconVersion)
+			}));
 			apps.sort(Core.compareApps)
 			return apps
 		}