+++ src/test/runTest.ts
@@ -27,6 +27,11 @@ async function main() {
 			 * See `code --help` for possible arguments.
 			 */
 			launchArgs: ['--disable-extensions'],
+			extensionTestsEnv: {
+				// Pass custom parameter into Mocha
+				MOCHA_OUTPUT_FILE: process.env.MOCHA_OUTPUT_FILE,
+			},
+
 		});
 	} catch (err) {
 		console.error(err);