+++ src/tree/App.js
@@ -1,9 +1,16 @@
+/* eslint-disable semi,@typescript-eslint/no-var-requires */
 const path = require('path')
 const EnhancedTreeItem = require('./EnhancedTreeItem')
+const { appsIconTempDir } = require('../temp-dir');
+const { getIconLocalPath } = require('../libs/app-icon');
 
 class App extends EnhancedTreeItem {
-	constructor(name, label, description, version, isPublic, isApproved, iconDir, theme, changes, iconVersion) {
-		super(label + (changes !== undefined ? (changes.length !== 0 ? ` ${EnhancedTreeItem.changedSymbol}` : "") : "") + (version > 1 ? ` (version ${version})` : ""))
+	constructor(name, label, description, version, isPublic, isApproved, theme, changes, iconVersion, isOpensource  = false) {
+		super(
+			label
+			+ (changes?.length !== 0 ? ` ${EnhancedTreeItem.changedSymbol}` : "")
+			+ (version > 1 ? ` (version ${version})` : "")
+		)
 		this.bareLabel = label
 		this.id = `${name}@${version}`
 		this.name = name
@@ -12,40 +19,40 @@ class App extends EnhancedTreeItem {
 		this.public = isPublic
 		this.approved = isApproved
 		this.level = 0
-		this.contextValue = "app" + (this.approved ? "_approved" : this.public ? "_public" : "") + (changes !== undefined ? (changes.length !== 0 ? "_changed" : "") : "")
+		this.contextValue = "app"
+			+ (isApproved ? "_approved" : "")
+			+ (!isApproved && isPublic ? "_public" : "")
+			+ (changes?.length !== 0 ? "_changed" : "");
 		this.theme = theme
-		this.iconVersion = iconVersion
+		/** Defines the final png icon filename. */
+		this.iconVersion = iconVersion;
 		this.changes = changes
+		/** Is the app example (formerly opensource)? If app is an example, then green square is not added to icon.  */
+		this.isOpensource = isOpensource;
 		this.tooltip = this.makeTooltip()
-		this.iconPath = this.makeIconPath(iconDir)
-		this.rawIcon = {
-			dark: path.join(iconDir, `${this.name}.${this.iconVersion}.png`),
-			light: path.join(iconDir, `${this.name}.${this.iconVersion}.dark.png`)
-		}
+		this.iconPath = this.makeIconPath();
+		this.rawIcon = getIconLocalPath(name, version, iconVersion, false);  // TODO here never used public icon. Why this exception?
 	}
-	makeIconPath(iconDir) {
-		if (!this.public) {
-			return {
-				dark: path.join(iconDir, `${this.name}.${this.iconVersion}.png`),
-				light: path.join(iconDir, `${this.name}.${this.iconVersion}.dark.png`)
-			}
-		}
-		else {
-			return {
-				dark: path.join(iconDir, `${this.name}.${this.iconVersion}.public.png`),
-				light: path.join(iconDir, `${this.name}.${this.iconVersion}.dark.public.png`)
-			}
-		}
+
+	/**
+	 * Get path to icon file.
+	 * If appp is public, then used public icon (it means icon with small green square in right bottom corner).
+	 * @param {string} iconDir Basedir of icons.
+	 * @returns {string}
+	 */
+	makeIconPath() {
+		return getIconLocalPath(this.name, this.version, this.iconVersion, this.public && !this.isOpensource);
 	}
 	makeTooltip() {
 		let tooltip = `${this.bareLabel}
 -----------------------
-Name: ${this.name}
+ID (Name): ${this.name}
 Description: ${this.description}
 Theme: ${this.theme}
 Version: ${this.version}
 Public: ${this.public}
-Approved: ${this.approved}`
+Approved: ${this.approved}`;
+
 		return tooltip
 	}
 }