+++ src/tree/EnhancedTreeItem.js
@@ -6,9 +6,10 @@ class EnhancedTreeItem extends vscode.TreeItem {
 		super(label, collapsibleState !== undefined ? collapsibleState : vscode.TreeItemCollapsibleState.Collapsed)
 	}
 
-    /**
-     * COMMON FUNCTIONS
-     */
+	/**
+	 * Return path to generic icon, which are placed statically in resouce dir.
+	 * @param {string} name Examples: "base", "code", "connection", "module, "rpc", "readme", "refresh", ...
+	 */
 	makeIconPath(name, inject = "") {
 		return {
 			dark: path.join(__dirname, '..', '..', "resources", "icons", "dark", `${name}${inject}.png`),